LPS-194418 Inline the component ObjectFieldSettingContributorRegistry in ObjectFieldLocalServiceImpl
